[Refactor] : Adapt MVVM architecture
