+++ DaeguGoodPriceShop/DaeguGoodPriceShop/View/MapViewController.swift
@@ -30,7 +30,7 @@ class MapViewController: UIViewController {
         configureView()
         configureMapView()
         configureBindings()
-        addAnnotation(shops: mapViewModel.model.shops)
+        addAnnotation()
     }
     
     private func configureView() {
@@ -71,7 +71,12 @@ class MapViewController: UIViewController {
         self.mapView.setRegion(region, animated: true)
     }
     
-    private func addAnnotation(shops: [Shop]) {
+    private func addAnnotation(category: ShopCategory? = nil) {
+        let shops = mapViewModel.getCategorizedShop(category: category)
+        addPins(shops: shops)
+    }
+    
+    private func addPins(shops: [Shop]) {
         for shop in shops {
             guard let category = shop.getCategory() else {
                 continue
@@ -81,54 +86,12 @@ class MapViewController: UIViewController {
         }
     }
     
-    private func removeAnnotation() {
-        mapView.removeAnnotations(mapView.annotations)
-    }
-    
     private func addPin(category: ShopCategory, coordinate: CLLocation) {
         let pin = ShopAnnotation(category: category, location: coordinate)
         mapView.addAnnotation(pin)
     }
-}
-
-extension MapViewController: MKMapViewDelegate {
-    func mapView(_ mapView: MKMapView, viewFor annotation: MKAnnotation) -> MKAnnotationView? {
-        guard let annotation = annotation as? ShopAnnotation else {
-            return nil
-        }
-        
-        var annotationView = self.mapView.dequeueReusableAnnotationView(withIdentifier: ShopAnnotationView.identifier)
-        
-        if annotationView == nil {
-            annotationView = MKAnnotationView(annotation: annotation, reuseIdentifier: ShopAnnotationView.identifier)
-            annotationView?.canShowCallout = false
-            annotationView?.contentMode = .scaleAspectFit
-        } else {
-            annotationView?.annotation = annotation
-        }
-        
-        let annotationImage: UIImage!
-        let size = CGSize(width: 30, height: 30)
-        UIGraphicsBeginImageContext(size)
-        
-        //TODO: Annotation Image 추가
-        switch annotation.category {
-        case .restaurant:
-            annotationImage = UIImage(systemName: "book")?.withTintColor(.red)
-        case .hair:
-            annotationImage = UIImage(systemName: "doc")?.withTintColor(.blue)
-        case .laundry:
-            annotationImage = UIImage(systemName: "paperplane")?.withTintColor(.orange)
-        case .beauty:
-            annotationImage = UIImage(systemName: "trash")?.withTintColor(.yellow)
-        case .bath:
-            annotationImage = UIImage(systemName: "swift")?.withTintColor(.purple)
-        }
-        
-        annotationImage.draw(in: CGRect(x: 0, y: 0, width: size.width, height: size.height))
-        let resizedImage = UIGraphicsGetImageFromCurrentImageContext()
-        annotationView?.image = resizedImage
-        
-        return annotationView
+    
+    private func removeAnnotation() {
+        mapView.removeAnnotations(mapView.annotations)
     }
 }
